app/requirements.txt
Django==3.2.18
djangorestframework==3.14.0
djangorestframework-simplejwt==4.6.3

psycopg2==2.9.9
Pillow==9.5.0